SCC-2324 Improve date and enumeration parsing
With feed back from research staff it was decided that the current string formats returned by this function were difficult to understand, especially for patrons who might not be familiar with standard data formats.

This updates the strings produced to be more "human-friendly". This mainly involves two things:

- Outputting full-text date strings instead of ISO-8601 date values. (e.g. Jan. 1, 2020 vs 2020-01-01)
- Adding Vol/No prefixes to enumeration values

Both of these updates should make these more recognizable for patrons, which is important as they are being given a place of much higher prominence on SCC as compared with the classic catalog