NOREF Remove pg gem from deployment
When deployed the pg gem is provided via a layer that is added to the lambda. At present the gem is being built as part of the deployment, which is causing issues by putting a newer version of the `pg` gem on the path ahead of the version intended for use with this function.