NOREF Fix terraform deployment (for real)
The previous fix did not resolve the issue because I did not understand the relationship between travis "stages". Each stage is provisioned it's own VM, meaning that cleaning the environment in one stage is not carried over to the next. (This seems ineffecient to me, but that's the way it works).

This moves the cleaning commands to each script for the deploy stages. This has the effect of duplicating some code, but as it's only two lines repeated twice, this doesn't seem too bad.